refactor(sdK): Move FuncASTTransform to graph.ast module to allow it to reused by future preprocessors
